+++ crates/core/src/proto_config.rs
@@ -124,10 +124,12 @@ derive_enum!(
 );
 
 derive_enum!(
-    #[derive(Copy, ConfigEnum)]
-    pub enum PinType {
+    #[derive(Copy, ConfigEnum, Default)]
+    #[cfg_attr(feature = "clap", derive(clap::ValueEnum))]
+    pub enum PinLocation {
         #[serde(alias = "store")]
         Global,
+        #[default]
         #[serde(alias = "cwd")]
         Local,
         #[serde(alias = "home")]
@@ -222,7 +224,7 @@ pub struct ProtoSettingsConfig {
     pub offline: ProtoOfflineConfig,
 
     #[setting(env = "PROTO_PIN_LATEST")]
-    pub pin_latest: Option<PinType>,
+    pub pin_latest: Option<PinLocation>,
 
     #[setting(default = true)]
     pub telemetry: bool,